fixed segfault when empty simple or double quotes

+++ includes/parsing.h
@@ -6,7 +6,7 @@
 /*   By: cafabre <cafabre@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/30 14:21:55 by willda-s          #+#    #+#             */
-/*   Updated: 2025/09/12 17:13:12 by cafabre          ###   ########.fr       */
+/*   Updated: 2025/09/12 19:28:44 by cafabre          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,7 +81,7 @@ typedef struct s_data
 	t_exec			*exec;
 	char			**envp;
 	int				errcode;
-	int				i;
+	size_t			i;
 }					t_data;
 
 typedef struct s_process_input
